file exist check

+++ generator_secret_json.py
@@ -1,15 +1,18 @@
-import json
-from django.core.management.utils import get_random_secret_key
+import os
 
-secret = {}
-secret['SECRET_KEY'] = get_random_secret_key()
-secret['DB_HOST'] = ''
-secret['DB_USER'] = ''
-secret['DB_DATABASE'] = ''
-secret['DB_PASSWORD'] = ''
+if not os.path.isfile('secret.json'):
+    import json
+    from django.core.management.utils import get_random_secret_key
 
-secret = json.dumps(secret)
+    secret = {}
+    secret['SECRET_KEY'] = get_random_secret_key()
+    secret['DB_HOST'] = ''
+    secret['DB_USER'] = ''
+    secret['DB_DATABASE'] = ''
+    secret['DB_PASSWORD'] = ''
 
-f = open('secret.json', 'w')
-f.write(secret)
-f.close()
+    secret = json.dumps(secret)
+
+    f = open('secret.json', 'w')
+    f.write(secret)
+    f.close()
